% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bd.R
\name{check_bd_flora}
\alias{check_bd_flora}
\alias{check_bd_pcob}
\alias{check_bd_trans}
\title{Chequeo Bases de Datos}
\usage{
check_bd_flora(x, mask = NULL)

check_bd_pcob(x, mask = NULL)

check_bd_trans(x, mask = NULL)
}
\arguments{
\item{x}{Base de datos}

\item{mask}{Capa mascara para filtrar las parcelas}
}
\value{
Devuelve alertas impresas en la consola o bien en modulos de shiny.
}
\description{
Chequeo Bases de Datos
}
