% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_carto_fun.R
\name{check_carto}
\alias{check_carto}
\title{Check cartography}
\usage{
check_carto(x, id)
}
\arguments{
\item{x}{Objeto sf.}

\item{id}{Nombre de la capa.}
}
\value{
Devuelve alguna alerta en la consola o en un modulo de shiny
}
\description{
Funcion para chequear la presencia y orden de los campos de la cartografia digital en base a los requerimientos tecnicos de CONAF.
}
