% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{check_n_area}
\alias{check_n_area}
\alias{check_n_rodal}
\alias{check_n_predio}
\title{Chequear N_Area}
\usage{
check_n_area(areas)

check_n_rodal(rodales)

check_n_predio(predios)
}
\arguments{
\item{areas}{Objeto sf de las areas de corta.}

\item{rodales}{Objeto sf con los rodales de las areas de corta.}

\item{predios}{Objeto sf con los predios de las areas de corta.}
}
\value{
reporte de errores.

reporte de errores.

reporte de errores.
}
\description{
función para chequear numeración de áreas preliminares

función para chequear numeración de rodales preliminares

función para chequear numeración de predios preliminares
}
