% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_appendices_utils.R
\name{cob_x_sp_fun}
\alias{cob_x_sp_fun}
\title{Cobertura por especie}
\usage{
cob_x_sp_fun(
  parcelas,
  bd_flora,
  field_parc = "N_Parc",
  cov_as_range = F,
  cov_fp = 1
)
}
\arguments{
\item{parcelas}{Nombre de parcelas.}

\item{bd_flora}{Base de datos con las parcelas.}

\item{field_parc}{Nombre del campo con ID de parcela.}

\item{cov_as_range}{logical. \code{TRUE} si desea dejar la cobertura de la flora como un rango. Valido solo para el PAS `151`.}

\item{cov_fp}{Cobertura para especies con COB_BB `fp` o no especificada. Elegir entre `0`, `0.5` o `1`. Default `1`.}
}
\value{
data.frame con cobertura de copas por especie.
}
\description{
Cobertura por especie
}
