% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_appendices_utils.R
\name{cup_coverage}
\alias{cup_coverage}
\title{Cobertura de copa}
\usage{
cup_coverage(x, y, method = "ellipse", na.value = 0)
}
\arguments{
\item{x, y}{Diametros de copa.}

\item{method}{Metodo; `circle` o `ellipse`. Default `circle`.}

\item{na.value}{Devuelve este valor si la cobertura es \code{NA}. Default `0`.}
}
\value{
Valor numerico con cobertura de copa.
}
\description{
Calcular cobertura de copas de un individuo.
}
\examples{
cup_coverage(x = 7, y = 3)
cup_coverage(x = 7, y = 3, method = "ellipse")
cup_coverage(x = NA, y = NA, na.value = 0)

}
