% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apendices.R
\name{apendice_2_3}
\alias{apendice_2_3}
\alias{apendice_5_PAS148}
\alias{apendice_5_PAS151}
\title{Apendices}
\usage{
apendice_2_3(
  PAS,
  bd_flora,
  bd_pcob = NULL,
  bd_trans = NULL,
  rodales,
  predios,
  portada = "default",
  provincia,
  huso = NULL
)

apendice_5_PAS148(
  bd_flora,
  rodales,
  tabla_predios,
  tabla_areas,
  tabla_attr_rodal,
  portada = "default",
  provincia,
  carto_uso_actual = NULL,
  obras = NULL,
  bd_fauna = NULL
)

apendice_5_PAS151(
  bd_flora,
  rodales,
  tabla_predios,
  tabla_areas,
  tabla_attr_rodal,
  portada = "default",
  provincia,
  obras = NULL
)
}
\arguments{
\item{PAS}{PAS correspondiente. Ingresar `148` o `151`.}

\item{bd_flora}{data.frame con base de datos de parcelas flristicas. Incluir si `add_parcelas` es \code{TRUE.}.}

\item{bd_pcob}{data.frame con base de datos de parcelas de cobertura.}

\item{rodales}{Objeto sf con capa de rodales.}

\item{predios}{Objeto sf con capa de predios.}

\item{portada}{Portada para los apendices. Disponible: `default` o `KIMAL`.}

\item{provincia}{Provincia. Character.}

\item{huso}{Huso. `18S` o `19S`.}

\item{tabla_predios}{data.frame con informacion de predios obtenida de la cartografia digital.}

\item{tabla_areas}{data.frame con informacion de areas de corta obtenida de la cartografia digital.}

\item{tabla_attr_rodal}{data.frame con tabla de atributos de rodal.}

\item{carto_uso_actual}{Objeto sf con capa de uso actual obtenida de la cartografia digital.}

\item{obras}{Opcional. Objeto sf con capa de rodales.}

\item{bd_fauna}{Opcional. data.frame con base de datos de fauna.}
}
\value{
Workbook Excel con apendices
}
\description{
Funciones para obtener apendices de los PAS
}
