% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_appendices.R
\name{get_appendices}
\alias{get_appendices}
\alias{apendice_2_3}
\alias{apendice_5_PAS148}
\alias{apendice_5_PAS151}
\alias{apendice_5_PAS151_nuevo}
\title{Apendices}
\usage{
apendice_2_3(
  PAS,
  bd_flora,
  bd_pcob = NULL,
  bd_trans = NULL,
  join_by = "coordenadas",
  predios,
  cov_as_range = F,
  cov_fp = 1,
  provincia,
  portada = "default",
  portada_opts = NULL
)

apendice_5_PAS148(
  bd_flora,
  rodales,
  tabla_predios,
  tabla_areas,
  tabla_attr_rodal,
  umbral_sp_est = 0.05,
  bd_flora_2 = NULL,
  provincia,
  portada = "default",
  portada_opts = NULL,
  carto_uso_actual = NULL,
  areas = NULL,
  obras = NULL
)

apendice_5_PAS151(
  bd_flora,
  rodales,
  tabla_predios,
  tabla_areas,
  tabla_attr_rodal,
  cov_as_range = F,
  cov_fp = 1,
  umbral_sp_est = 0.05,
  bd_flora_2 = NULL,
  portada = "default",
  portada_opts = NULL,
  provincia,
  areas = NULL,
  obras = NULL
)

apendice_5_PAS151_nuevo(
  bd_flora,
  rodales,
  tabla_predios,
  tabla_areas,
  tabla_attr_rodal,
  cov_as_range = F,
  cov_fp = 1,
  umbral_sp_est = 0.05,
  bd_flora_2 = NULL,
  portada = "default",
  portada_opts = NULL,
  provincia,
  areas = NULL,
  obras = NULL
)
}
\arguments{
\item{PAS}{PAS correspondiente. Ingresar `148` o `151`.}

\item{bd_flora}{data.frame con base de datos de parcelas flristicas. Incluir si `add_parcelas` es \code{TRUE.}. ver `detail`.}

\item{bd_pcob}{data.frame con base de datos de parcelas de cobertura.}

\item{bd_trans}{data.frame con base de datos de transectos.}

\item{join_by}{Union de las bases de datos `parcela` o `coordenadas`.}

\item{predios}{Objeto sf con capa de predios.}

\item{cov_as_range}{logical. \code{TRUE} si desea dejar la cobertura de la flora como un rango. Valido solo para el PAS `151`.}

\item{cov_fp}{Cobertura para especies con Cob_BB `fp` o no especificada. Elegir entre `0`, `0.5` o `1`. Default `1`.}

\item{provincia}{Provincia. Character.}

\item{portada}{Portada para los apendices. Disponible: `default`, `MLP612`, `KIM753` u `otra`.}

\item{portada_opts}{Opciones para personalizar una portada. ver `details`.}

\item{rodales}{Objeto sf con capa de rodales.}

\item{tabla_predios}{data.frame con informacion de predios obtenida de la cartografia digital.}

\item{tabla_areas}{data.frame con informacion de areas de corta obtenida de la cartografia digital.}

\item{tabla_attr_rodal}{data.frame con tabla de atributos de rodal. ver `details`.}

\item{umbral_sp_est}{Porcentaje de representividad minimo de una especie cuya densidad se va a estimar. default \code{0.05}.}

\item{bd_flora_2}{Opcional. Base de datos de flora que contenga las parcelas utilizadas para atributar areas de corta sin parcelas directa o cercana. ver `details`.}

\item{carto_uso_actual}{Objeto sf con capa de uso actual obtenida de la cartografia digital.}

\item{areas}{Objeto sf con capa de areas de corta. default \code{NULL}.}

\item{obras}{Opcional. Objeto sf con capa de rodales.}
}
\value{
Workbook Excel con apendices
}
\description{
Funciones para obtener apendices de los PAS
}
\details{
Antes de ingresar alguna de las bases de datos, se recomienda verificar que estas contengan los campos requeridos, entre otros errores.
Para esto puede utilizar las funciones [check_bd_flora()], [check_bd_pcob()], [check_bd_trans()].
Una ayuda a preparar las bases de datos podrian ser las funciones [prepare_bd_flora()], [prepare_bd_pcob()] o [prepare_bd_trans()].

En el parametro `portada_opts` debe ingresar una lista con los argumentos de la funcion [portada_opts()].

El data.frame debe contener el mismo formato que el arrojado con la funcion [get_tabla_attr_rodal()].

La `bd_flora_2` debe contener al menos los campos; "Parcela", "Especie", "Cob_BB", "Nha" y "Habito".
}
