% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_appendices_utils.R
\name{italic_sp}
\alias{italic_sp}
\title{Especies en cursiva}
\usage{
italic_sp(x)
}
\arguments{
\item{x}{Tipo vegetacional.}
}
\value{
Mismo texto con `*` para dar formato cursiva.
}
\description{
Especies en cursiva
}
\examples{
italic_sp(x = c(
"Bosque de Schinus latifolius-Quillaja saponaria",
"Matorral con suculentas de Flourensia thurifera",
"Plantación de especies nativas de Acacia caven-Quillaja saponaria"))

}
