% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_appendices_utils.R
\name{nha_x_sp_fun}
\alias{nha_x_sp_fun}
\title{Densidad por especie}
\usage{
nha_x_sp_fun(parcelas, bd, field_parc = "N_Parc", add_var = NULL)
}
\arguments{
\item{parcelas}{Nombre de parcelas.}

\item{bd}{Base de datos con las parcelas. Debe contener el Nha por especie y parcela de muestreo.}

\item{field_parc}{Nombre del campo con ID de parcela.}

\item{add_var}{Variable/s que desea agregar. default \code{NULL}.}
}
\value{
data.frame con densidad (ind/ha) por especie.
}
\description{
Densidad por especie
}
