% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_appendices_utils.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Cobertura de la parcela}
\usage{
plot_coverage(x, plot_size = 500, percent = FALSE, digits = 1L, ...)
}
\arguments{
\item{x}{Vector numerico de coberturas individuales de copa.}

\item{plot_size}{Superficie de la parcela en m2. Default \code{500}.}

\item{percent}{Logico; si es \code{TRUE} agrega el simbolo de porcentaje.}

\item{digits}{Cantidad de decimales.}

\item{...}{Otros argumentos de la funcion base [formatC()].}
}
\value{
Cobertura de parcela formato texto.
}
\description{
Calcular cobertura de copas de la parcela.
}
\examples{
plot_coverage(c(20, 15), percent = FALSE, digits = 2, decimal.mark = ",")
plot_coverage(c(NA, NA, 20, 15), percent = TRUE, digits = 1, decimal.mark = ",")

}
