% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portadas.R
\name{portada}
\alias{portada}
\alias{wb_portada_PAS148y151}
\alias{portada_opts}
\title{Portadas de apendices}
\usage{
wb_portada_PAS148y151(wb, PAS, apendice, provincia, opts = portada_opts())

portada_opts(
  tipo_proj = "EIA",
  nom_proj = NULL,
  logo = NULL,
  plantilla = "default"
)
}
\arguments{
\item{wb}{Workbook.}

\item{PAS}{PAS (Permiso ambiental sectorial) correspondiente. \code{148} o \code{151}.}

\item{apendice}{Numero de apendice. \code{2}= Densidad de parcelas; \code{3}= Ubicacion de parcelas; \code{5}= Tablas formulario.}

\item{provincia}{Nombre de la provincia.}

\item{opts}{Opciones para personalizar una portada. usar [portada_opts()].}

\item{tipo_proj}{Tipo de proyecto. `DIA` o `EIA`. default `EIA`.}

\item{nom_proj}{Nombre del proyecto.}

\item{logo}{Ruta del archivo de tipo imagen con el logo del cliente.}

\item{plantilla}{Plantilla realizada para proyectos en particular, disponible `KIM753`, `MLP612` y `default`.}
}
\value{
Workbook con portada.

Workbook con portada.
}
\description{
Portadas de apendices

Ajuste de portada
}
