% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{prepare_bd}
\alias{prepare_bd}
\alias{prepare_bd_flora}
\alias{prepare_bd_pcob}
\alias{prepare_bd_trans}
\title{Prepare BD inputs}
\usage{
prepare_bd_flora(BD, rodales, PAS, cut_by_rod, include_fp = FALSE)

prepare_bd_pcob(BD)

prepare_bd_trans(BD)
}
\arguments{
\item{BD}{data.frame o ruta con la base de datos}

\item{rodales}{objeto sf con los rodales}

\item{PAS}{PAS correspondiente. Ingresar \code{148}, \code{149} o \code{151}.}

\item{cut_by_rod}{logical. \code{TRUE} para seleccionar parcelas dentro de rodales.}

\item{include_fp}{Incluir especies fuera de parcela. Incluye ademas especies cuya `Cob_BB` sea \code{NA} o \code{---}.}
}
\value{
Inputs formateados
}
\description{
Funciones para preparar formato de los datos ingresados.
}
