% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inputs.R
\name{prepare_bd_flora}
\alias{prepare_bd_flora}
\title{Prepare inputs}
\usage{
prepare_bd_flora(BD, rodales, PAS, cut_by_rod)
}
\arguments{
\item{BD}{data.frame o ruta con la base de datos}

\item{rodales}{objeto sf con los rodales}

\item{cut_by_rod}{logical. \code{TRUE} para seleccionar parcelas dentro de rodales}
}
\value{
Inputs formateados
}
\description{
Prepare inputs
}
