% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carto_digital.R
\name{carto_digital}
\alias{carto_digital}
\alias{cart_rodales}
\alias{cart_area}
\alias{cart_suelos}
\alias{cart_rang_pend}
\alias{cart_predios}
\alias{cart_parcelas}
\alias{cart_uso_actual}
\alias{cart_hidro}
\alias{cart_hidro_osm}
\alias{cart_caminos}
\alias{cart_caminos_osm}
\alias{cart_curv_niv}
\alias{get_carto_digital}
\title{Cartografia digital}
\usage{
cart_rodales(PAS, rodales, from_RCA = F, RCA = NULL, dec_sup = 2L)

cart_area(PAS, areas, dec_sup = 2L, from_RCA = F, RCA = NULL)

cart_suelos(
  PAS,
  areas,
  suelos_LB = NULL,
  dec_sup = 2,
  from_RCA = F,
  RCA = NULL
)

cart_rang_pend(PAS, areas, dem, stat_slope = "median", dec_sup = 2)

cart_predios(predios, cut_by_prov = NULL, dec_sup = 2)

cart_parcelas(PAS, bd_flora, rodales, cut_by_rod, include_fp)

cart_uso_actual(predios, catastro, suelos, dec_sup = 2)

cart_hidro(
  predios,
  fuente_hidro = c("MOP", "BCN"),
  cut = c("clip", "buffer", "crop", "crop_by_row"),
  buffer = 0
)

cart_hidro_osm(
  predios,
  cut = c("clip", "buffer", "crop", "crop_by_row"),
  buffer = 0
)

cart_caminos(
  predios,
  cut = c("clip", "buffer", "crop", "crop_by_row"),
  buffer = 0
)

cart_caminos_osm(
  predios,
  cut = c("clip", "buffer", "crop", "crop_by_row"),
  buffer = 0
)

cart_curv_niv(
  predios,
  dem,
  cut = c("clip", "buffer", "crop", "crop_by_row"),
  buffer = 0,
  step = 10
)

get_carto_digital(
  PAS,
  areas,
  rodales,
  predios,
  dem,
  stat_slope = "median",
  suelos_LB = NULL,
  cut_by_prov = NULL,
  add_parcelas = FALSE,
  bd_flora = NULL,
  cut_by_rod = F,
  include_fp = FALSE,
  from_RCA = FALSE,
  RCA = NULL,
  add_uso_actual = FALSE,
  catastro = NULL,
  suelos = NULL,
  add_caminos = FALSE,
  add_caminos_osm = FALSE,
  caminos_arg = list(cut = "clip", buffer = 0),
  add_hidro = FALSE,
  fuente_hidro = c("MOP", "BCN"),
  add_hidro_osm = FALSE,
  hidro_arg = list(cut = "clip", buffer = 0),
  add_curv_niv = FALSE,
  curv_niv_arg = list(cut = "clip", buffer = 0),
  step = 10,
  dec_sup = 2
)
}
\arguments{
\item{PAS}{PAS correspondiente. Ingresar \code{148}, \code{149} o \code{151}.}

\item{rodales}{Objeto sf con los rodales de las areas de corta.}

\item{from_RCA}{Logico; \code{TRUE} si la capa probiene de una RCA aprobada. Default \code{FALSE}.}

\item{RCA}{Numero de la RCA aprobada. Icluir si `from_RCA` es \code{TRUE}. Default \code{NULL}.}

\item{dec_sup}{Cantidad de decimales para la superficie en hectareas. Default `2`.}

\item{areas}{Objeto sf de las areas de corta.}

\item{suelos_LB}{Objeto sf con la informacion de suelo de línea de base para áreas de corta.}

\item{dem}{Objeto SpatRaster o ruta del archivo raster.}

\item{stat_slope}{Resumir pendiente por 'mean' o 'median'.}

\item{predios}{Objeto sf con los predios de las areas de corta.}

\item{cut_by_prov}{Cortar predios por provincia.}

\item{bd_flora}{Base de datos de flora. Incluir si `add_parcelas` es \code{TRUE}.}

\item{cut_by_rod}{Seleccionar parcelas dentro de rodales.}

\item{include_fp}{Incluir especies fuera de parcela. Incluye ademas especies cuya `Cob_BB` sea \code{NA} o \code{---}.}

\item{catastro}{Objeto sf del catastro de CONAF.}

\item{suelos}{Objeto sf con informacion de suelo para los predios.}

\item{fuente_hidro}{Fuente de la capa hidrografica. `MOP` o `BCN`.}

\item{cut}{Tipo de corte para aplicar a la capa. Opciones: `clip`, `buffer`, `crop`, `crop_by_row`.}

\item{buffer}{Distancia (m) de buffer para aplicar el corte de la capa.}

\item{step}{Intervalo para curvas de nivel. Incluir si `add_curv_niv` es \code{TRUE}. Default `10`.}

\item{add_parcelas}{Logico; \code{TRUE} para incluir capa de parcelas.}

\item{add_uso_actual}{Logico; \code{TRUE} para incluir capa de uso actual.}

\item{add_caminos}{Logico; \code{TRUE} para incluir capa de caminos.}

\item{add_caminos_osm}{Logico; \code{TRUE} para incluir capa de caminos desde OpenStreetMap.}

\item{caminos_arg}{Lista de argumentos para cortar capa de caminos. Default `cut`= 'clip'; `buffer` = 0.}

\item{add_hidro}{Logico; \code{TRUE} para incluir capa de hidrografia.}

\item{add_hidro_osm}{Logico; \code{TRUE} para incluir capa de hidrografia desde OpenStreetMap.}

\item{hidro_arg}{Lista de argumentos para cortar capa de hidrografia Default `cut`= 'clip'; `buffer` = 0.}

\item{add_curv_niv}{Logico; \code{TRUE} para incluir capa de curvas de nivel.}

\item{curv_niv_arg}{Lista de argumentos para cortar capa de curvas de nivel. Default `cut`= 'clip'; `buffer` = 0.}
}
\value{
Capas vectoriales de la cartografia digital
}
\description{
Funciones diversas para elaborar la cartografia digital
}
