% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carto_digital.R
\name{carto_digital}
\alias{carto_digital}
\alias{cart_rodales}
\alias{cart_area}
\alias{cart_suelos}
\alias{cart_rang_pend}
\alias{cart_predios}
\alias{cart_parcelas}
\alias{cart_uso_actual}
\alias{cart_hidro}
\alias{cart_hidro_osm}
\alias{cart_caminos}
\alias{cart_caminos_osm}
\alias{cart_curv_niv}
\alias{get_carto_digital}
\title{Cartografia digital}
\usage{
cart_rodales(
  PAS,
  rodales,
  TipoFor_num = NULL,
  from_RCA = F,
  RCA = NULL,
  dec_sup = 2
)

cart_area(PAS, areas, dec_sup = 2, from_RCA = F, RCA = NULL)

cart_suelos(PAS, areas, dec_sup = 2, from_RCA = F, RCA = NULL)

cart_rang_pend(PAS, areas, dem, dec_sup = 2)

cart_predios(predios, cut_by_prov = NULL, dec_sup = 2)

cart_parcelas(PAS, bd_flora, rodales, cut_by_rod)

cart_uso_actual(predios, catastro, suelos, dec_sup = 2)

cart_hidro(predios, fuente_hidro, cut = "clip", buffer = 0)

cart_hidro_osm(predios, cut = "clip", buffer = 0)

cart_caminos(predios, cut = "clip", buffer = 0)

cart_caminos_osm(predios, cut = "clip", buffer = 0)

cart_curv_niv(predios, dem, cut = "clip", buffer = 0, step = 10)

get_carto_digital(
  PAS = 148,
  areas,
  rodales,
  cut_by_rod,
  TipoFor_num = T,
  predios,
  cut_by_prov,
  dem,
  add_parcelas = F,
  bd_flora = NULL,
  from_RCA = F,
  RCA = NULL,
  add_uso_actual = F,
  catastro = NULL,
  suelos = NULL,
  add_caminos = F,
  add_caminos_osm = F,
  caminos_arg = list(cut = "clip", buffer = 0),
  add_hidro = F,
  fuente_hidro = NULL,
  add_hidro_osm = F,
  hidro_arg = list(cut = "clip", buffer = 0),
  add_curv_niv = F,
  curv_niv_arg = list(cut = "clip", buffer = 0),
  step = 10,
  dec_sup = 2
)
}
\arguments{
\item{PAS}{PAS correspondiente. Ingresar `148` o `151`.}

\item{rodales}{Objeto sf con los rodales de las areas de corta.}

\item{TipoFor_num}{Logico; si es \code{TRUE} declara que el campo `Tipo_For` esta con el codigo numerico.}

\item{from_RCA}{Logico; \code{TRUE} si la capa probiene de una RCA aprobada. Default \code{FALSE}.}

\item{RCA}{Numero de la RCA aprobada. Icluir si `from_RCA` es \code{TRUE}. Default \code{NULL}.}

\item{dec_sup}{Cantidad de decimales para la superficie en hectareas. Default `2`.}

\item{areas}{Objeto sf de las areas de corta.}

\item{dem}{Objeto SpatRaster o ruta del archivo raster.}

\item{predios}{Objeto sf con los predios de las areas de corta.}

\item{bd_flora}{Base de datos de flora. Incluir si `add_parcelas` es \code{TRUE.}.}

\item{catastro}{Objeto sf del catastro de CONAF.}

\item{suelos}{Objeto sf con informacion de suelo.}

\item{fuente_hidro}{Fuente de la capa hidrografica. `MOP` o `BCN`.}

\item{cut}{Tipo de corte para aplicar a la capa. Opciones: `clip`, `buffer`, `crop`, `crop_by_row`.}

\item{buffer}{Distancia (m) de buffer para aplicar el corte de la capa.}

\item{step}{Intervalo para curvas de nivel. Incluir si `add_curv_niv` es \code{TRUE}. Default `10`.}

\item{add_parcelas}{Logico; \code{TRUE} para incluir capa de parcelas.}

\item{add_uso_actual}{Logico; \code{TRUE} para incluir capa de uso actual.}

\item{add_caminos}{Logico; \code{TRUE} para incluir capa de caminos.}

\item{add_caminos_osm}{Logico; \code{TRUE} para incluir capa de caminos desde OpenStreetMap.}

\item{caminos_arg}{Lista de argumentos para cortar capa de caminos. Default `cut`= 'clip'; `buffer` = 0.}

\item{add_hidro}{Logico; \code{TRUE} para incluir capa de hidrografia.}

\item{add_hidro_osm}{Logico; \code{TRUE} para incluir capa de hidrografia desde OpenStreetMap.}

\item{hidro_arg}{Lista de argumentos para cortar capa de hidrografia Default `cut`= 'clip'; `buffer` = 0.}

\item{add_curv_niv}{Logico; \code{TRUE} para incluir capa de curvas de nivel.}

\item{curv_niv_arg}{Lista de argumentos para cortar capa de curvas de nivel. Default `cut`= 'clip'; `buffer` = 0.}
}
\value{
Capas vectoriales de la cartografia digital
}
\description{
Funciones diversas para elaborar la cartografia digital
}
