% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bd.R
\name{check_bd_flora}
\alias{check_bd_flora}
\title{Chequeo Bases de Datos}
\usage{
check_bd_flora(x, y = NULL, shinyalert = F)
}
\arguments{
\item{x}{Base de datos de flora}

\item{y}{Capa}

\item{shinyalert}{Logico; usar \code{TRUE} en una shiny app. Default \code{FALSE}}
}
\value{
Devuelve alertas impresas en la consola o bien en modulos de shiny.
}
\description{
Chequeo Bases de Datos
}
