% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_apendices_utils_fun.R
\name{cob_x_sp_fun}
\alias{cob_x_sp_fun}
\title{Cobertura por especie}
\usage{
cob_x_sp_fun(parcelas, bd_flora)
}
\arguments{
\item{parcelas}{Nombre de parcelas.}

\item{bd_flora}{Base de datos con las parcelas.}
}
\value{
data.frame con cobertura de copas por especie.
}
\description{
Cobertura por especie
}
