% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pred_rod_area.R
\name{get_pred_rod_area}
\alias{get_pred_rod_area}
\title{get_pred_rod_area}
\usage{
get_pred_rod_area(
  PAS,
  LB,
  obras,
  predios,
  suelos,
  group_by_LB = NULL,
  sep_by_soil = TRUE,
  group_by_dist = FALSE,
  distance_max = ifelse(group_by_dist == F, NULL, 50),
  cut_by_prov = FALSE,
  provincia = NULL,
  n_rodal_ord = FALSE,
  orden_rodal = "NS-OE",
  dec_sup = 2L
)
}
\arguments{
\item{PAS}{PAS correspondiente. Ingresar `148` o `151`.}

\item{LB}{Objeto sf con la linea base de flora.}

\item{obras}{Objeto sf con las obras del proyecto.}

\item{predios}{Objeto sf con los limites ´prediales.}

\item{suelos}{Objeto sf con la informacion de suelo.}

\item{group_by_LB}{lista de campos por los que agrupar los rodales de linea base.}

\item{sep_by_soil}{Logico; si es \code{TRUE}, separa las areas de corta por capa de suelo.}

\item{group_by_dist}{Logico; si es \code{TRUE}, agrupa las areas de corta que estan a cierta distancia.}

\item{distance_max}{Distancia maxima para agrupar los poligonos. Utilizar si `group_by_dist` es \code{TRUE}.}

\item{cut_by_prov}{Logico; si es \code{TRUE}, genera las capas para la provincia indicada}

\item{provincia}{Nombre de la provincia. Utilizar si `cut_by_prov` es \code{TRUE}.}

\item{n_rodal_ord}{Logico; si es \code{TRUE}, ordena espacialmente los rodales, de lo contrario sigue la numeración del numero de poligono.}

\item{orden_rodal}{Orden espacial para los rodales. Se aplica cuando `n_rodal_ord` es \code{TRUE}. Mas detalle en la documentacion de [st_order()].}

\item{dec_sup}{Cantidad de decimales para la superficie en hectareas. Default `2`.}
}
\value{
Lista con los sf de Predios, Rodales y Areas de corta
}
\description{
Función para obtener de forma preliminar capa de Predios, Rodales y Areas de corta
}
