% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tabla_attr_rodal.R
\name{get_tabla_attr_rodal}
\alias{get_tabla_attr_rodal}
\title{Tabla atributos de rodal}
\usage{
get_tabla_attr_rodal(PAS, bd_flora, rodales_def)
}
\arguments{
\item{PAS}{PAS correspondiente. Ingresar `148` o `151`.}

\item{bd_flora}{Base de datos de parcelas dentro de los rodales.}

\item{rodales_def}{Objeto sf con poligono de rodales.}
}
\value{
Tabla con los atributos de rodal
}
\description{
Tabla atributos de rodal
}
