% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_apendices_utils_fun.R
\name{nha_x_sp_fun}
\alias{nha_x_sp_fun}
\title{Densidad por especie}
\usage{
nha_x_sp_fun(parcelas, bd, add_var = NULL)
}
\arguments{
\item{parcelas}{Nombre de parcelas.}

\item{bd}{Base de datos con las parcelas.}

\item{add_var}{`NULL`; Variable/s que desea agregar.}
}
\value{
data.frame con densidad (ind/ha) por especie.
}
\description{
Densidad por especie
}
