% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_functions.R
\name{group_by_distance}
\alias{group_by_distance}
\alias{my_union}
\alias{get_slope}
\alias{st_order}
\title{Funciones Geoespaciales}
\usage{
group_by_distance(x, distance)

my_union(x, y)

get_slope(dem, x)

st_order(x, order = "NS-OE")
}
\arguments{
\item{x, y}{Objeto sf.}

\item{distance}{Distancia maxima para agrupar.}

\item{dem}{Digital Elevation Model. Puede ser un objeto SpatRaster o bien la ruta de la imagen.}

\item{order}{Orden espacial.}
}
\value{
Objetos sf o numeros
}
\description{
`group_by_distance` crea agrupaciones en base a la distancia ingresada.
`my_union` regresa la union de dos capas, se crucen o no las geometrias.
`get_slope` regresa la pendiente media.
`st_order` crea un orden espacial del objeto sf ingresado.
}
